WIP: implementing Finance services

+++ Isotainer.Module.Finance/Core/ViewModels/GeneralCost/GeneralCostUpdateResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace Isotainer.Module.Finance.Core.ViewModels.GeneralCost;
 
-public record GeneralCostUpdateResponse(Guid GeneralCostId, decimal Cost);
+public record GeneralCostUpdateResponse(Guid GeneralCostId, double Cost);
